updated command
