+++ SoftLayer/fixtures/SoftLayer_Account.py
@@ -704,17 +704,192 @@ getVirtualDedicatedRacks = [{
     'metricTrackingObjectId': 10,
 }]
 
-getDedicatedHosts = [{
-    'datacenter': {
-        'name': 'dal05'
-    },
-    'memoryCapacity': 242,
-    'name': 'test-dedicated',
-    'diskCapacity': 1200,
-    'guestCount': 1,
-    'cpuCount': 56,
-    'id': 12345
-}]
+getDedicatedHosts = [
+    {
+        "cpuCount": 56,
+        "createDate": "2021-11-18T15:13:57-06:00",
+        "diskCapacity": 1200,
+        "id": 656700,
+        "memoryCapacity": 242,
+        "name": "dedicatedhost01",
+        "guestCount": 0,
+        "allocationStatus": {
+            "cpuAllocated": 0,
+            "cpuAvailable": 56,
+            "cpuCount": 56,
+            "diskAllocated": 0,
+            "diskAvailable": 1200,
+            "diskCapacity": 1200,
+            "guestCount": 0,
+            "memoryAllocated": 0,
+            "memoryAvailable": 242,
+            "memoryCapacity": 242
+        },
+        "backendRouter": {
+            "accountId": 1,
+            "bareMetalInstanceFlag": 0,
+            "domain": "softlayer.com",
+            "fullyQualifiedDomainName": "bcr01a.dal13.softlayer.com",
+            "hardwareStatusId": 5,
+            "hostname": "bcr01a.dal13",
+            "id": 1883692,
+            "notes": "",
+            "provisionDate": None,
+            "serviceProviderId": 1,
+            "serviceProviderResourceId": None
+        },
+        "billingItem": {
+            "allowCancellationFlag": 1,
+            "cancellationDate": None,
+            "categoryCode": "dedicated_virtual_hosts",
+            "createDate": "2021-11-18T15:13:59-06:00",
+            "currentHourlyCharge": "0",
+            "cycleStartDate": "2023-01-04T00:07:36-06:00",
+            "description": "56 Cores X 242 RAM X 1.2 TB",
+            "hostName": "dedicatedhost01",
+            "hourlyRecurringFee": "0",
+            "hoursUsed": "465",
+            "id": 910104016,
+            "laborFee": "0",
+            "laborFeeTaxRate": "0",
+            "lastBillDate": "2023-01-04T00:07:36-06:00",
+            "modifyDate": "2023-01-04T00:07:36-06:00",
+            "nextBillDate": "2023-02-04T00:00:00-06:00",
+            "oneTimeFee": "0",
+            "oneTimeFeeTaxRate": "0",
+            "orderItemId": 874370304,
+            "parentId": None,
+            "recurringFee": "0",
+            "recurringFeeTaxRate": "0",
+            "recurringMonths": 1,
+            "serviceProviderId": 1,
+            "setupFee": "0",
+            "setupFeeTaxRate": "0",
+            "orderItem": {
+                "categoryCode": "dedicated_virtual_hosts",
+                "description": "56 Cores X 242 RAM X 1.2 TB",
+                "hostName": "dedicatedhost01",
+                "hourlyRecurringFee": "0",
+                "id": 874370304,
+                "itemId": 10195,
+                "itemPriceId": "200269",
+                "laborAfterTaxAmount": "0",
+                "laborFee": "0",
+                "laborFeeTaxRate": "0",
+                "laborTaxAmount": "0",
+                "oneTimeAfterTaxAmount": "0",
+                "oneTimeFee": "0",
+                "oneTimeFeeTaxRate": "0",
+                "oneTimeTaxAmount": "0",
+                "parentId": None,
+                "presetId": None,
+                "promoCodeId": None,
+                "quantity": None,
+                "recurringAfterTaxAmount": "0",
+                "recurringFee": "0",
+                "recurringTaxAmount": "0",
+                "setupAfterTaxAmount": "0",
+                "setupFee": "0",
+                "setupFeeDeferralMonths": None,
+                "setupFeeTaxRate": "0",
+                "setupTaxAmount": "0",
+                "order": {
+                    "accountId": 307608,
+                    "createDate": "2021-11-18T15:13:57-06:00",
+                    "id": 85857762,
+                    "impersonatingUserRecordId": None,
+                    "modifyDate": "2021-11-18T15:14:00-06:00",
+                    "orderQuoteId": None,
+                    "orderTypeId": 11,
+                    "presaleEventId": None,
+                    "privateCloudOrderFlag": False,
+                    "status": "APPROVED",
+                    "userRecordId": 7650493,
+                    "userRecord": {
+                        "accountId": 307608,
+                        "address1": "4849 Alpha Rd",
+                        "city": "Dallas",
+                        "companyName": "SoftLayer Internal - Development Community",
+                        "country": "US",
+                        "createDate": "2019-06-10T15:21:33-07:00",
+                        "daylightSavingsTimeFlag": False,
+                        "denyAllResourceAccessOnCreateFlag": False,
+                        "displayName": "danielc",
+                        "email": "daniel.cabero@jalasoft.com",
+                        "firstName": "daniel",
+                        "forumPasswordHash": "unused",
+                        "iamAuthorizationStatus": 1,
+                        "iamId": "IBMid-550003CUGT",
+                        "id": 7650493,
+                        "isMasterUserFlag": False,
+                        "lastName": "cabero",
+                        "localeId": 1,
+                        "managedByFederationFlag": False,
+                        "managedByOpenIdConnectFlag": True,
+                        "minimumPasswordLifeHours": 0,
+                        "modifyDate": "2019-12-13T07:29:07-06:00",
+                        "openIdConnectUserName": "daniel.cabero@jalasoft.com",
+                        "parentId": 167758,
+                        "passwordExpireDate": None,
+                        "permissionCheckLikeMasterUserFlag": 0,
+                        "postalCode": "75244-4608",
+                        "pptpVpnAllowedFlag": False,
+                        "preventPreviousPasswords": 0,
+                        "savedId": "7650493",
+                        "secondaryLoginManagementFlag": True,
+                        "secondaryLoginRequiredFlag": None,
+                        "secondaryPasswordModifyDate": "2019-06-10T15:21:33-07:00",
+                        "secondaryPasswordTimeoutDays": 0,
+                        "sslVpnAllowedFlag": False,
+                        "state": "TX",
+                        "statusDate": None,
+                        "timezoneId": 113,
+                        "userStatusId": 1001,
+                        "username": "sl307608-dcabero",
+                        "vpnManualConfig": False,
+                        "ibmIdLink": {
+                            "createDate": "2019-06-10T15:21:34-07:00",
+                            "defaultFlag": 1,
+                            "destinationUserAlphanumericId": "daniel%2ecabero@jalasoft%2ecom",
+                            "destinationUserId": None,
+                            "iamIdVerificationFlag": 2,
+                            "id": 1051135,
+                            "realm": "IBMid",
+                            "serviceProviderId": 348,
+                            "uniqueIdentifier": "IBMid-550003CUGT",
+                            "userId": 7650493
+                        },
+                        "locale": {
+                            "friendlyName": "English",
+                            "id": 1,
+                            "languageTag": "en-US",
+                            "name": "English"
+                        },
+                        "timezone": {
+                            "id": 113,
+                            "longName": "(GMT-06:00) America/Chicago - CST",
+                            "name": "America/Chicago",
+                            "offset": "-0600",
+                            "shortName": "CST"
+                        },
+                        "userStatus": {
+                            "id": 1001,
+                            "keyName": "ACTIVE",
+                            "name": "Active"
+                        }
+                    }
+                }
+            },
+            "resourceTableId": 656700
+        },
+        "datacenter": {
+            "id": 1854895,
+            "longName": "Dallas 13",
+            "name": "dal13",
+            "statusId": 2
+        }
+    }
+]
 
 getUsers = [
     {'displayName': 'ChristopherG',
@@ -1294,190 +1469,3 @@ getNetworkGateways = [{
         'keyName': 'ACTIVE'
     }
 }]
-
-getFilteredDedicatedHosts = [
-    {
-        "cpuCount": 56,
-        "createDate": "2021-11-18T15:13:57-06:00",
-        "diskCapacity": 1200,
-        "id": 656700,
-        "memoryCapacity": 242,
-        "name": "dedicatedhost01",
-        "guestCount": 0,
-        "allocationStatus": {
-            "cpuAllocated": 0,
-            "cpuAvailable": 56,
-            "cpuCount": 56,
-            "diskAllocated": 0,
-            "diskAvailable": 1200,
-            "diskCapacity": 1200,
-            "guestCount": 0,
-            "memoryAllocated": 0,
-            "memoryAvailable": 242,
-            "memoryCapacity": 242
-        },
-        "backendRouter": {
-            "accountId": 1,
-            "bareMetalInstanceFlag": 0,
-            "domain": "softlayer.com",
-            "fullyQualifiedDomainName": "bcr01a.dal13.softlayer.com",
-            "hardwareStatusId": 5,
-            "hostname": "bcr01a.dal13",
-            "id": 1883692,
-            "notes": "",
-            "provisionDate": None,
-            "serviceProviderId": 1,
-            "serviceProviderResourceId": None
-        },
-        "billingItem": {
-            "allowCancellationFlag": 1,
-            "cancellationDate": None,
-            "categoryCode": "dedicated_virtual_hosts",
-            "createDate": "2021-11-18T15:13:59-06:00",
-            "currentHourlyCharge": "0",
-            "cycleStartDate": "2023-01-04T00:07:36-06:00",
-            "description": "56 Cores X 242 RAM X 1.2 TB",
-            "hostName": "dedicatedhost01",
-            "hourlyRecurringFee": "0",
-            "hoursUsed": "465",
-            "id": 910104016,
-            "laborFee": "0",
-            "laborFeeTaxRate": "0",
-            "lastBillDate": "2023-01-04T00:07:36-06:00",
-            "modifyDate": "2023-01-04T00:07:36-06:00",
-            "nextBillDate": "2023-02-04T00:00:00-06:00",
-            "oneTimeFee": "0",
-            "oneTimeFeeTaxRate": "0",
-            "orderItemId": 874370304,
-            "parentId": None,
-            "recurringFee": "0",
-            "recurringFeeTaxRate": "0",
-            "recurringMonths": 1,
-            "serviceProviderId": 1,
-            "setupFee": "0",
-            "setupFeeTaxRate": "0",
-            "orderItem": {
-                "categoryCode": "dedicated_virtual_hosts",
-                "description": "56 Cores X 242 RAM X 1.2 TB",
-                "hostName": "dedicatedhost01",
-                "hourlyRecurringFee": "0",
-                "id": 874370304,
-                "itemId": 10195,
-                "itemPriceId": "200269",
-                "laborAfterTaxAmount": "0",
-                "laborFee": "0",
-                "laborFeeTaxRate": "0",
-                "laborTaxAmount": "0",
-                "oneTimeAfterTaxAmount": "0",
-                "oneTimeFee": "0",
-                "oneTimeFeeTaxRate": "0",
-                "oneTimeTaxAmount": "0",
-                "parentId": None,
-                "presetId": None,
-                "promoCodeId": None,
-                "quantity": None,
-                "recurringAfterTaxAmount": "0",
-                "recurringFee": "0",
-                "recurringTaxAmount": "0",
-                "setupAfterTaxAmount": "0",
-                "setupFee": "0",
-                "setupFeeDeferralMonths": None,
-                "setupFeeTaxRate": "0",
-                "setupTaxAmount": "0",
-                "order": {
-                    "accountId": 307608,
-                    "createDate": "2021-11-18T15:13:57-06:00",
-                    "id": 85857762,
-                    "impersonatingUserRecordId": None,
-                    "modifyDate": "2021-11-18T15:14:00-06:00",
-                    "orderQuoteId": None,
-                    "orderTypeId": 11,
-                    "presaleEventId": None,
-                    "privateCloudOrderFlag": False,
-                    "status": "APPROVED",
-                    "userRecordId": 7650493,
-                    "userRecord": {
-                        "accountId": 307608,
-                        "address1": "4849 Alpha Rd",
-                        "city": "Dallas",
-                        "companyName": "SoftLayer Internal - Development Community",
-                        "country": "US",
-                        "createDate": "2019-06-10T15:21:33-07:00",
-                        "daylightSavingsTimeFlag": False,
-                        "denyAllResourceAccessOnCreateFlag": False,
-                        "displayName": "danielc",
-                        "email": "daniel.cabero@jalasoft.com",
-                        "firstName": "daniel",
-                        "forumPasswordHash": "unused",
-                        "iamAuthorizationStatus": 1,
-                        "iamId": "IBMid-550003CUGT",
-                        "id": 7650493,
-                        "isMasterUserFlag": False,
-                        "lastName": "cabero",
-                        "localeId": 1,
-                        "managedByFederationFlag": False,
-                        "managedByOpenIdConnectFlag": True,
-                        "minimumPasswordLifeHours": 0,
-                        "modifyDate": "2019-12-13T07:29:07-06:00",
-                        "openIdConnectUserName": "daniel.cabero@jalasoft.com",
-                        "parentId": 167758,
-                        "passwordExpireDate": None,
-                        "permissionCheckLikeMasterUserFlag": 0,
-                        "postalCode": "75244-4608",
-                        "pptpVpnAllowedFlag": False,
-                        "preventPreviousPasswords": 0,
-                        "savedId": "7650493",
-                        "secondaryLoginManagementFlag": True,
-                        "secondaryLoginRequiredFlag": None,
-                        "secondaryPasswordModifyDate": "2019-06-10T15:21:33-07:00",
-                        "secondaryPasswordTimeoutDays": 0,
-                        "sslVpnAllowedFlag": False,
-                        "state": "TX",
-                        "statusDate": None,
-                        "timezoneId": 113,
-                        "userStatusId": 1001,
-                        "username": "sl307608-dcabero",
-                        "vpnManualConfig": False,
-                        "ibmIdLink": {
-                            "createDate": "2019-06-10T15:21:34-07:00",
-                            "defaultFlag": 1,
-                            "destinationUserAlphanumericId": "daniel%2ecabero@jalasoft%2ecom",
-                            "destinationUserId": None,
-                            "iamIdVerificationFlag": 2,
-                            "id": 1051135,
-                            "realm": "IBMid",
-                            "serviceProviderId": 348,
-                            "uniqueIdentifier": "IBMid-550003CUGT",
-                            "userId": 7650493
-                        },
-                        "locale": {
-                            "friendlyName": "English",
-                            "id": 1,
-                            "languageTag": "en-US",
-                            "name": "English"
-                        },
-                        "timezone": {
-                            "id": 113,
-                            "longName": "(GMT-06:00) America/Chicago - CST",
-                            "name": "America/Chicago",
-                            "offset": "-0600",
-                            "shortName": "CST"
-                        },
-                        "userStatus": {
-                            "id": 1001,
-                            "keyName": "ACTIVE",
-                            "name": "Active"
-                        }
-                    }
-                }
-            },
-            "resourceTableId": 656700
-        },
-        "datacenter": {
-            "id": 1854895,
-            "longName": "Dallas 13",
-            "name": "dal13",
-            "statusId": 2
-        }
-    }
-]
